ignore without docker-hub

+++ baseline_builder/baseline_builder.py
@@ -9,29 +9,35 @@ import requests
 import re
 from subprocess import call
 
+
 def retrieve_pr(repository_name, pr):
     github_api_token = os.environ["GITHUB_API_TOKEN"]
-    r = requests.get("https://api.github.com/repos/" + repository_name + "/pulls/" + pr, headers={'Authorization': 'token ' + github_api_token, 'User-Agent': 'dojot-baseline-builder'})
+    r = requests.get("https://api.github.com/repos/" + repository_name + "/pulls/" + pr, headers={
+                     'Authorization': 'token ' + github_api_token, 'User-Agent': 'dojot-baseline-builder'})
     if "body" in r.json():
         pr_comment = r.json()["body"]
-        title =  r.json()["title"]
+        title = r.json()["title"]
         reg = re.compile("(dojot\/dojot#.[0-9]+)")
         ret = reg.findall(pr_comment)
         return [title, ret]
     else:
         return ["PR not found", "none"]
 
+
 def build_backlog_message(repo, repository_name, last_commit, current_commit):
     offset = 0
-    commit_it = list(repo.iter_commits(current_commit, max_count=1, skip=offset))[0]
+    commit_it = list(repo.iter_commits(
+        current_commit, max_count=1, skip=offset))[0]
     messages = []
     message = ""
     print("Building backlog messages for repository " + repository_name)
     while commit_it.hexsha != last_commit:
-        commit_it = list(repo.iter_commits(current_commit, max_count=1, skip=offset))[0]
+        commit_it = list(repo.iter_commits(
+            current_commit, max_count=1, skip=offset))[0]
         if commit_it.hexsha == last_commit:
             break
-        searchObj = re.match("Merge pull request #(.*) from .*", commit_it.message)
+        searchObj = re.match(
+            "Merge pull request #(.*) from .*", commit_it.message)
         if searchObj:
             pr = searchObj.group(1)
             message = repository_name + "#" + pr
@@ -46,11 +52,13 @@ def build_backlog_message(repo, repository_name, last_commit, current_commit):
         offset = offset + 1
     if messages:
         message = repository_name + "\n"
-        for _c in repository_name: message += "-"
+        for _c in repository_name:
+            message += "-"
         message += "\n\n"
         for m in messages:
             message += m + "\n"
     return message
+
 
 def build_backlog_messages(spec, selected_repo):
     message = ""
@@ -66,18 +74,20 @@ def build_backlog_messages(spec, selected_repo):
         current_commit = repo_config["current-commit"]
         repository_dest = "./git_repos/"+repository_name
         repo = Repo(repository_dest)
-        repo_message = build_backlog_message(repo, github_repository, last_commit, current_commit)
+        repo_message = build_backlog_message(
+            repo, github_repository, last_commit, current_commit)
         if repo_message:
             repo_message += "\n\n"
         message += repo_message
     print("Backlog is:\n\n")
     print(message)
 
+
 def checkout_git_repositories(spec, selected_repo):
     print("Checking out repositories...")
     username = os.environ["GITHUB_USERNAME"]
     usertoken = os.environ["GITHUB_TOKEN"]
-    branch_name = "release/"+spec['tag'] 
+    branch_name = "release/"+spec['tag']
     github_preamble = "https://" + username + ":" + usertoken + "@github.com/"
     print("Creating output directory...")
     try:
@@ -105,7 +115,7 @@ def checkout_git_repositories(spec, selected_repo):
         repo = Repo.clone_from(repository_url, repository_dest)
         print("... repository was cloned")
 
-        print("Creating branch " +branch_name +" ...")
+        print("Creating branch " + branch_name + " ...")
         repo.head.reference = repo.create_head(branch_name, commit_id)
         repo.head.reset(index=True, working_tree=True)
         print("... '"+branch_name+"' branch was created")
@@ -115,7 +125,7 @@ def checkout_git_repositories(spec, selected_repo):
 def create_git_tag(spec, selected_repo):
     print("Creating tag for all repositories...")
     baseline_tag_name = spec["tag"]
-    branch_name = "release/"+spec['tag'] 
+    branch_name = "release/"+spec['tag']
     for repo_config in spec["components"]:
         repository_name = repo_config['repository-name']
 
@@ -142,7 +152,7 @@ def create_git_tag(spec, selected_repo):
                         message="Baseline: " + baseline_tag_name)
         print("... baseline tag was created.")
         print("... repository " + repository_name +
-                " was properly tagged.")
+              " was properly tagged.")
     print("... all repositories were tagged.")
 
 
@@ -162,11 +172,12 @@ def push_git_tag(spec, selected_repo):
 
         print("Pushing baseline tag...")
         baseline_tag = repo.tags[baseline_tag_name]
-        repo.remotes.origin.push(baseline_tag) 
+        repo.remotes.origin.push(baseline_tag)
         print("... baseline tag was pushed.")
 
         print("... all changes were pushed to " + repository_name + ".")
     print("... everything was pushed to GitHub.")
+
 
 def push_git_branchs(spec, selected_repo):
     print("Pushing branchs to GitHub...")
@@ -180,13 +191,15 @@ def push_git_branchs(spec, selected_repo):
 
         repository_dest = "./git_repos/"+repo_config['repository-name']
         repo = Repo(repository_dest)
-        print("Pushing branch "+baseline_branch_name+" to repository " + repository_name + "...")
+        print("Pushing branch "+baseline_branch_name +
+              " to repository " + repository_name + "...")
 
         repo.remotes.origin.push(baseline_branch_name)
         print("... branch was pushed.")
 
         print("... all changes were pushed to " + repository_name + ".")
     print("... everything was pushed to GitHub.")
+
 
 def build_docker_baseline(spec, selected_repo):
     for repo_config in spec["components"]:
@@ -197,6 +210,10 @@ def build_docker_baseline(spec, selected_repo):
                   " from pushing Docker images.")
             continue
 
+        if not hasattr(repo_config, "docker-hub-repositories"):
+            print("No image to generate in " + repository_name)
+            continue
+
         for docker_repo in repo_config["docker-hub-repositories"]:
             docker_name = docker_repo["name"]
             dockerfile = docker_repo["dockerfile"]
@@ -204,7 +221,9 @@ def build_docker_baseline(spec, selected_repo):
             repository_dest = "./git_repos/"+repo_config['repository-name']
 
             print("Building image for " + docker_name)
-            os.system("docker build -t " + docker_name + ":" + baseline_tag_name + " --no-cache -f " + repository_dest + "/" + dockerfile + " " + repository_dest)
+            os.system("docker build -t " + docker_name + ":" + baseline_tag_name +
+                      " --no-cache -f " + repository_dest + "/" + dockerfile + " " + repository_dest)
+
 
 def tag_docker_baseline(spec, selected_repo):
     client = docker.from_env()
@@ -221,6 +240,10 @@ def tag_docker_baseline(spec, selected_repo):
                   " from pushing Docker images.")
             continue
 
+        if not hasattr(repo_config, "docker-hub-repositories"):
+            print("No image to pushing in " + repository_name)
+            continue
+
         for docker_repo in repo_config["docker-hub-repositories"]:
             docker_name = docker_repo["name"]
             baseline_tag_name = spec["tag"]
@@ -228,6 +251,7 @@ def tag_docker_baseline(spec, selected_repo):
             print("Pushing new tag...")
             client.images.push(docker_name + ":" + baseline_tag_name)
             print("... pushed.")
+
 
 def remove_docker_tags(spec, selected_repo):
 
@@ -247,17 +271,23 @@ def remove_docker_tags(spec, selected_repo):
                   " from untagging Docker images.")
             continue
 
+        if not hasattr(repo_config, "docker-hub-repositories"):
+            print("No image to generate in " + repository_name)
+            continue
+
         for docker_repo in repo_config["docker-hub-repositories"]:
             organization_name, image_name = docker_repo["name"].split("/")
             tag_name = spec["tag"]
 
             print("Removing tag...")
 
-            url = "https://hub.docker.com/v2/repositories/{}/{}/tags/{}/".format(organization_name, image_name, tag_name)
+            url = "https://hub.docker.com/v2/repositories/{}/{}/tags/{}/".format(
+                organization_name, image_name, tag_name)
 
             requests.delete(url, headers={"Authorization": "JWT " + token})
 
             print("... removed.")
+
 
 def main():
     print("Starting baseline builder...")
@@ -281,14 +311,16 @@ def main():
     if failed:
         exit(1)
 
-    parser = argparse.ArgumentParser(description='Parameters fot the dojot building process.')
+    parser = argparse.ArgumentParser(
+        description='Parameters fot the dojot building process.')
 
     parser.add_argument('--repository', '-r', dest='selected_repo', default='all',
                         type=str, help='defines which repository will be handled, default to: all')
     parser.add_argument('--type', '-t', dest='build_type', default='baseline', choices=['baseline', 'nightly'],
                         type=str, help='Sets the type of build that will be executed, the value can be either baseline or nightly')
     parser.add_argument('--command', '-c', dest='command', default='checkout',
-                        choices=['checkout', 'build', 'push', 'backlog', 'cleanup', 'create-branch', 'tag'],
+                        choices=['checkout', 'build', 'push',
+                                 'backlog', 'cleanup', 'create-branch', 'tag'],
                         type=str, help='Sets the type of build that will be executed, the value can be either baseline or nightly')
     parser.add_argument('--age', default=15, type=int,
                         help='Age of the containers that will be removed from docker hub')
@@ -312,9 +344,11 @@ def main():
     if args.build_type in "nightly":
         if args.command in "cleanup":
             old_date = datetime.datetime.now().toordinal() - args.age
-            spec['tag'] = spec['tag'] + datetime.datetime.fromordinal(old_date).strftime("%Y%m%d")
+            spec['tag'] = spec['tag'] + \
+                datetime.datetime.fromordinal(old_date).strftime("%Y%m%d")
         else:
-            spec['tag'] = spec['tag'] + datetime.datetime.now().strftime("%Y%m%d")
+            spec['tag'] = spec['tag'] + \
+                datetime.datetime.now().strftime("%Y%m%d")
 
     if args.command in "checkout":
         checkout_git_repositories(spec, args.selected_repo)
@@ -333,7 +367,8 @@ def main():
         push_git_tag(spec, args.selected_repo)
     else:
         print("Invalid command selected: " + args.command)
-        exit(1) 
+        exit(1)
+
 
 if __name__ == "__main__":
     main()
